python/requirements.txt
matplotlib
numpy
torch
pyserial
openai